requirements.txt
fastapi
uvicorn
pillow
python-dotenv
requests
pandas
pyarrow
openpyxl
